refactor(store): move to plugins directory
