+++ frontend/src/main.js
@@ -2,7 +2,7 @@
 
 import {createApp} from 'vue';
 import App from './App.vue';
-import store from './store/index';
+import store from './plugins/store';
 import VueAxios from 'vue-axios';
 import router from './plugins/router';
 import i18n from './i18n';
